Move filesystem helper to specific file
